# net/tsaddr: use bart in NewContainsIPFunc, add tests, benchmarks