--- a/net/tsaddr/tsaddr.go
+++ b/net/tsaddr/tsaddr.go
@@ -11,6 +11,7 @@ import (
 	"slices"
 	"sync"
 
+	"github.com/gaissmai/bart"
 	"go4.org/netipx"
 	"tailscale.com/net/netaddr"
 	"tailscale.com/types/views"
@@ -165,6 +166,10 @@ func FalseContainsIPFunc() func(ip netip.Addr) bool {
 	return func(ip netip.Addr) bool { return false }
 }
 
+// pathForTest is a test hook for NewContainsIPFunc, to test that it took the
+// right construction path.
+var pathForTest = func(string) {}
+
 // NewContainsIPFunc returns a func that reports whether ip is in addrs.
 //
 // It's optimized for the cases of addrs being empty and addrs
@@ -176,32 +181,50 @@ func NewContainsIPFunc(addrs views.Slice[netip.Prefix]) func(ip netip.Addr) bool
 	// Specialize the three common cases: no address, just IPv4
 	// (or just IPv6), and both IPv4 and IPv6.
 	if addrs.Len() == 0 {
+		pathForTest("empty")
 		return func(netip.Addr) bool { return false }
 	}
-	// If any addr is more than a single IP, then just do the slow
-	// linear thing until
-	// https://github.com/inetaf/netaddr/issues/139 is done.
+	// If any addr is a prefix with more than a single IP, then do either a
+	// linear scan or a bart table, depending on the number of addrs.
 	if addrs.ContainsFunc(func(p netip.Prefix) bool { return !p.IsSingleIP() }) {
-		acopy := addrs.AsSlice()
-		return func(ip netip.Addr) bool {
-			for _, a := range acopy {
-				if a.Contains(ip) {
-					return true
-				}
+		if addrs.Len() > 6 {
+			pathForTest("bart")
+			// Built a bart table.
+			t := &bart.Table[struct{}]{}
+			for i := range addrs.Len() {
+				t.Insert(addrs.At(i), struct{}{})
 			}
-			return false
+			return func(ip netip.Addr) bool {
+				_, ok := t.Get(ip)
+				return ok
+			}
+		} else {
+			pathForTest("linear-contains")
+			// Small enough to do a linear search.
+			acopy := addrs.AsSlice()
+			return func(ip netip.Addr) bool {
+				for _, a := range acopy {
+					if a.Contains(ip) {
+						return true
+					}
+				}
+				return false
+			}
 		}
 	}
 	// Fast paths for 1 and 2 IPs:
 	if addrs.Len() == 1 {
+		pathForTest("one-ip")
 		a := addrs.At(0)
 		return func(ip netip.Addr) bool { return ip == a.Addr() }
 	}
 	if addrs.Len() == 2 {
+		pathForTest("two-ip")
 		a, b := addrs.At(0), addrs.At(1)
 		return func(ip netip.Addr) bool { return ip == a.Addr() || ip == b.Addr() }
 	}
 	// General case:
+	pathForTest("ip-map")
 	m := map[netip.Addr]bool{}
 	for i := range addrs.Len() {
 		m[addrs.At(i).Addr()] = true