… cmd/derper/depaware.txt
… cmd/stund/depaware.txt
… cmd/tailscale/depaware.txt